libct/cg: move GetAllPids out of utils.go
This is just moving the code around to ease the code review, no other
changes.
